+++ docs/en/src/general-config.md
@@ -5,7 +5,7 @@ This configuration is exposed via the `xplr.config.general` API.
 For now, kindly refer to [**init.lua**][1]
 
 > **NEED HELP:** Auto generate rest of the docs from [src/init.lua][1]
-> using [docs/script/generate.lua][2].
+> using [docs/script/generate.py][2].
 
 [1]: https://github.com/sayanarijit/xplr/blob/main/src/init.lua
-[2]: https://github.com/sayanarijit/xplr/blob/main/docs/script/generate.lua
+[2]: https://github.com/sayanarijit/xplr/blob/main/docs/script/generate.py
