+++ docs/en/src/input-operation.md
@@ -1,0 +1,25 @@
+# Input Operation
+
+Cursor based input operation can be one of the following:
+
+- { SetCursor = int }
+- { InsertCharacter = str }
+- "GoToPreviousCharacter"
+- "GoToNextCharacter"
+- "GoToPreviousWord"
+- "GoToNextWord"
+- "GoToStart"
+- "GoToEnd"
+- "DeletePreviousCharacter"
+- "DeleteNextCharacter"
+- "DeletePreviousWord"
+- "DeleteNextWord"
+- "DeleteLine"
+
+## Also See:
+
+- [Message][1]
+- [Full List of Messages][2]
+
+[1]: message.md
+[2]: messages.md
