+++ src/app.rs
@@ -1,1163 +1,37 @@
 use crate::config::Config;
 use crate::config::Mode;
+pub use crate::directory_buffer::DirectoryBuffer;
 use crate::explorer;
 use crate::input::{InputOperation, Key};
 use crate::lua;
-use crate::permissions::Permissions;
+pub use crate::msg::in_::external::Command;
+pub use crate::msg::in_::external::ExplorerConfig;
+pub use crate::msg::in_::external::NodeFilter;
+pub use crate::msg::in_::external::NodeFilterApplicable;
+pub use crate::msg::in_::external::NodeSorter;
+pub use crate::msg::in_::external::NodeSorterApplicable;
+pub use crate::msg::in_::ExternalMsg;
+pub use crate::msg::in_::InternalMsg;
+pub use crate::msg::in_::MsgIn;
+pub use crate::msg::out::MsgOut;
+pub use crate::node::Node;
+pub use crate::node::ResolvedNode;
+pub use crate::pipe::Pipe;
 use crate::ui::Layout;
 use anyhow::{bail, Result};
 use chrono::{DateTime, Local};
-use humansize::{file_size_opts as options, FileSize};
 use indexmap::set::IndexSet;
 use serde::{Deserialize, Serialize};
-use std::cmp::Ordering;
 use std::collections::HashMap;
 use std::collections::VecDeque;
 use std::env;
 use std::fs;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use tui_input::{Input, InputRequest};
 
 pub const VERSION: &str = env!("CARGO_PKG_VERSION");
 pub const TEMPLATE_TABLE_ROW: &str = "TEMPLATE_TABLE_ROW";
 pub const UNSUPPORTED_STR: &str = "???";
-
-fn to_humansize(size: u64) -> String {
-    size.file_size(options::CONVENTIONAL)
-        .unwrap_or_else(|_| format!("{} B", size))
-}
-
-fn mime_essence(path: &Path, is_dir: bool) -> String {
-    if is_dir {
-        String::from("inode/directory")
-    } else {
-        mime_guess::from_path(&path)
-            .first()
-            .map(|m| m.essence_str().to_string())
-            .unwrap_or_default()
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Pipe {
-    pub path: String,
-    pub msg_in: String,
-    pub selection_out: String,
-    pub result_out: String,
-    pub directory_nodes_out: String,
-    pub global_help_menu_out: String,
-    pub logs_out: String,
-    pub history_out: String,
-}
-
-impl Pipe {
-    fn from_session_path(path: &str) -> Result<Self> {
-        let path = PathBuf::from(path).join("pipe");
-
-        let msg_in = path.join("msg_in").to_string_lossy().to_string();
-
-        let selection_out =
-            path.join("selection_out").to_string_lossy().to_string();
-
-        let result_out = path.join("result_out").to_string_lossy().to_string();
-
-        let directory_nodes_out = path
-            .join("directory_nodes_out")
-            .to_string_lossy()
-            .to_string();
-
-        let global_help_menu_out = path
-            .join("global_help_menu_out")
-            .to_string_lossy()
-            .to_string();
-
-        let logs_out = path.join("logs_out").to_string_lossy().to_string();
-
-        let history_out =
-            path.join("history_out").to_string_lossy().to_string();
-
-        Ok(Self {
-            path: path.to_string_lossy().to_string(),
-            msg_in,
-            selection_out,
-            result_out,
-            directory_nodes_out,
-            global_help_menu_out,
-            logs_out,
-            history_out,
-        })
-    }
-}
-
-#[derive(Debug, Clone, Eq, Hash, PartialEq, Serialize, Deserialize)]
-pub struct ResolvedNode {
-    pub absolute_path: String,
-    pub extension: String,
-    pub is_dir: bool,
-    pub is_file: bool,
-    pub is_readonly: bool,
-    pub mime_essence: String,
-    pub size: u64,
-    pub human_size: String,
-}
-
-impl ResolvedNode {
-    pub fn from(path: PathBuf) -> Self {
-        let extension = path
-            .extension()
-            .map(|e| e.to_string_lossy().to_string())
-            .unwrap_or_default();
-
-        let (is_dir, is_file, is_readonly, size) = path
-            .metadata()
-            .map(|m| {
-                (m.is_dir(), m.is_file(), m.permissions().readonly(), m.len())
-            })
-            .unwrap_or((false, false, false, 0));
-
-        let mime_essence = mime_essence(&path, is_dir);
-        let human_size = to_humansize(size);
-
-        Self {
-            absolute_path: path.to_string_lossy().to_string(),
-            extension,
-            is_dir,
-            is_file,
-            is_readonly,
-            mime_essence,
-            size,
-            human_size,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Eq, Hash, PartialEq, Serialize, Deserialize)]
-pub struct Node {
-    pub parent: String,
-    pub relative_path: String,
-    pub absolute_path: String,
-    pub extension: String,
-    pub is_dir: bool,
-    pub is_file: bool,
-    pub is_symlink: bool,
-    pub is_broken: bool,
-    pub is_readonly: bool,
-    pub mime_essence: String,
-    pub size: u64,
-    pub human_size: String,
-    pub permissions: Permissions,
-    pub canonical: Option<ResolvedNode>,
-    pub symlink: Option<ResolvedNode>,
-}
-
-impl Node {
-    pub fn new(parent: String, relative_path: String) -> Self {
-        let absolute_path = PathBuf::from(&parent)
-            .join(&relative_path)
-            .to_string_lossy()
-            .to_string();
-
-        let path = PathBuf::from(&absolute_path);
-
-        let extension = path
-            .extension()
-            .map(|e| e.to_string_lossy().to_string())
-            .unwrap_or_default();
-
-        let (is_broken, maybe_canonical_meta) = path
-            .canonicalize()
-            .map(|p| (false, Some(ResolvedNode::from(p))))
-            .unwrap_or_else(|_| (true, None));
-
-        let (is_symlink, is_dir, is_file, is_readonly, size, permissions) =
-            path.symlink_metadata()
-                .map(|m| {
-                    (
-                        m.file_type().is_symlink(),
-                        m.is_dir(),
-                        m.is_file(),
-                        m.permissions().readonly(),
-                        m.len(),
-                        Permissions::from(&m),
-                    )
-                })
-                .unwrap_or_else(|_| {
-                    (false, false, false, false, 0, Permissions::default())
-                });
-
-        let mime_essence = mime_essence(&path, is_dir);
-        let human_size = to_humansize(size);
-
-        Self {
-            parent,
-            relative_path,
-            absolute_path,
-            extension,
-            is_dir,
-            is_file,
-            is_symlink,
-            is_broken,
-            is_readonly,
-            mime_essence,
-            size,
-            human_size,
-            permissions,
-            canonical: maybe_canonical_meta.clone(),
-            symlink: if is_symlink {
-                maybe_canonical_meta
-            } else {
-                None
-            },
-        }
-    }
-}
-
-impl Ord for Node {
-    fn cmp(&self, other: &Self) -> Ordering {
-        // Notice that the we flip the ordering on costs.
-        // In case of a tie we compare positions - this step is necessary
-        // to make implementations of `PartialEq` and `Ord` consistent.
-        other.relative_path.cmp(&self.relative_path)
-    }
-}
-
-impl PartialOrd for Node {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub struct DirectoryBuffer {
-    pub parent: String,
-    pub nodes: Vec<Node>,
-    pub total: usize,
-    pub focus: usize,
-}
-
-impl DirectoryBuffer {
-    pub fn new(parent: String, nodes: Vec<Node>, focus: usize) -> Self {
-        let total = nodes.len();
-        Self {
-            parent,
-            nodes,
-            total,
-            focus,
-        }
-    }
-
-    pub fn focused_node(&self) -> Option<&Node> {
-        self.nodes.get(self.focus)
-    }
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub enum InternalMsg {
-    AddLastFocus(String, Option<String>),
-    SetDirectory(DirectoryBuffer),
-    HandleKey(Key),
-}
-
-#[derive(Debug, Clone, Copy, Hash, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub enum NodeSorter {
-    ByRelativePath,
-    ByIRelativePath,
-    ByExtension,
-    ByIsDir,
-    ByIsFile,
-    ByIsSymlink,
-    ByIsBroken,
-    ByIsReadonly,
-    ByMimeEssence,
-    BySize,
-
-    ByCanonicalAbsolutePath,
-    ByICanonicalAbsolutePath,
-    ByCanonicalExtension,
-    ByCanonicalIsDir,
-    ByCanonicalIsFile,
-    ByCanonicalIsReadonly,
-    ByCanonicalMimeEssence,
-    ByCanonicalSize,
-
-    BySymlinkAbsolutePath,
-    ByISymlinkAbsolutePath,
-    BySymlinkExtension,
-    BySymlinkIsDir,
-    BySymlinkIsFile,
-    BySymlinkIsReadonly,
-    BySymlinkMimeEssence,
-    BySymlinkSize,
-}
-
-#[derive(Debug, Clone, Eq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct NodeSorterApplicable {
-    pub sorter: NodeSorter,
-    #[serde(default)]
-    pub reverse: bool,
-}
-
-impl PartialEq for NodeSorterApplicable {
-    fn eq(&self, other: &NodeSorterApplicable) -> bool {
-        self.sorter == other.sorter
-    }
-}
-
-impl std::hash::Hash for NodeSorterApplicable {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        self.sorter.hash(state);
-    }
-}
-
-impl NodeSorterApplicable {
-    fn reversed(mut self) -> Self {
-        self.reverse = !self.reverse;
-        self
-    }
-
-    fn apply(&self, a: &Node, b: &Node) -> Ordering {
-        let order = match self.sorter {
-            NodeSorter::ByRelativePath => {
-                natord::compare(&a.relative_path, &b.relative_path)
-            }
-            NodeSorter::ByIRelativePath => {
-                natord::compare_ignore_case(&a.relative_path, &b.relative_path)
-            }
-            NodeSorter::ByExtension => a.extension.cmp(&b.extension),
-            NodeSorter::ByIsDir => a.is_dir.cmp(&b.is_dir),
-            NodeSorter::ByIsFile => a.is_file.cmp(&b.is_file),
-            NodeSorter::ByIsSymlink => a.is_symlink.cmp(&b.is_symlink),
-            NodeSorter::ByIsBroken => a.is_broken.cmp(&b.is_broken),
-            NodeSorter::ByIsReadonly => a.is_readonly.cmp(&b.is_readonly),
-            NodeSorter::ByMimeEssence => a.mime_essence.cmp(&b.mime_essence),
-            NodeSorter::BySize => a.size.cmp(&b.size),
-
-            NodeSorter::ByCanonicalAbsolutePath => natord::compare(
-                &a.canonical
-                    .as_ref()
-                    .map(|s| s.absolute_path.clone())
-                    .unwrap_or_default(),
-                &b.canonical
-                    .as_ref()
-                    .map(|s| s.absolute_path.clone())
-                    .unwrap_or_default(),
-            ),
-
-            NodeSorter::ByICanonicalAbsolutePath => {
-                natord::compare_ignore_case(
-                    &a.canonical
-                        .as_ref()
-                        .map(|s| s.absolute_path.clone())
-                        .unwrap_or_default(),
-                    &b.canonical
-                        .as_ref()
-                        .map(|s| s.absolute_path.clone())
-                        .unwrap_or_default(),
-                )
-            }
-
-            NodeSorter::ByCanonicalExtension => a
-                .canonical
-                .as_ref()
-                .map(|s| &s.extension)
-                .cmp(&b.canonical.as_ref().map(|s| &s.extension)),
-
-            NodeSorter::ByCanonicalIsDir => a
-                .canonical
-                .as_ref()
-                .map(|s| &s.is_dir)
-                .cmp(&b.canonical.as_ref().map(|s| &s.is_dir)),
-
-            NodeSorter::ByCanonicalIsFile => a
-                .canonical
-                .as_ref()
-                .map(|s| &s.is_file)
-                .cmp(&b.canonical.as_ref().map(|s| &s.is_file)),
-
-            NodeSorter::ByCanonicalIsReadonly => a
-                .canonical
-                .as_ref()
-                .map(|s| &s.is_readonly)
-                .cmp(&b.canonical.as_ref().map(|s| &s.is_readonly)),
-
-            NodeSorter::ByCanonicalMimeEssence => a
-                .canonical
-                .as_ref()
-                .map(|s| &s.mime_essence)
-                .cmp(&b.canonical.as_ref().map(|s| &s.mime_essence)),
-
-            NodeSorter::ByCanonicalSize => a
-                .canonical
-                .as_ref()
-                .map(|s| &s.size)
-                .cmp(&b.canonical.as_ref().map(|s| &s.size)),
-
-            NodeSorter::BySymlinkAbsolutePath => natord::compare(
-                &a.symlink
-                    .as_ref()
-                    .map(|s| s.absolute_path.clone())
-                    .unwrap_or_default(),
-                &b.symlink
-                    .as_ref()
-                    .map(|s| s.absolute_path.clone())
-                    .unwrap_or_default(),
-            ),
-
-            NodeSorter::ByISymlinkAbsolutePath => natord::compare_ignore_case(
-                &a.symlink
-                    .as_ref()
-                    .map(|s| s.absolute_path.clone())
-                    .unwrap_or_default(),
-                &b.symlink
-                    .as_ref()
-                    .map(|s| s.absolute_path.clone())
-                    .unwrap_or_default(),
-            ),
-
-            NodeSorter::BySymlinkExtension => a
-                .symlink
-                .as_ref()
-                .map(|s| &s.extension)
-                .cmp(&b.symlink.as_ref().map(|s| &s.extension)),
-
-            NodeSorter::BySymlinkIsDir => a
-                .symlink
-                .as_ref()
-                .map(|s| &s.is_dir)
-                .cmp(&b.symlink.as_ref().map(|s| &s.is_dir)),
-
-            NodeSorter::BySymlinkIsFile => a
-                .symlink
-                .as_ref()
-                .map(|s| &s.is_file)
-                .cmp(&b.symlink.as_ref().map(|s| &s.is_file)),
-
-            NodeSorter::BySymlinkIsReadonly => a
-                .symlink
-                .as_ref()
-                .map(|s| &s.is_readonly)
-                .cmp(&b.symlink.as_ref().map(|s| &s.is_readonly)),
-
-            NodeSorter::BySymlinkMimeEssence => a
-                .symlink
-                .as_ref()
-                .map(|s| &s.mime_essence)
-                .cmp(&b.symlink.as_ref().map(|s| &s.mime_essence)),
-
-            NodeSorter::BySymlinkSize => a
-                .symlink
-                .as_ref()
-                .map(|s| &s.size)
-                .cmp(&b.symlink.as_ref().map(|s| &s.size)),
-        };
-        if self.reverse {
-            order.reverse()
-        } else {
-            order
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, Hash, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub enum NodeFilter {
-    RelativePathIs,
-    RelativePathIsNot,
-
-    IRelativePathIs,
-    IRelativePathIsNot,
-
-    RelativePathDoesStartWith,
-    RelativePathDoesNotStartWith,
-
-    IRelativePathDoesStartWith,
-    IRelativePathDoesNotStartWith,
-
-    RelativePathDoesContain,
-    RelativePathDoesNotContain,
-
-    IRelativePathDoesContain,
-    IRelativePathDoesNotContain,
-
-    RelativePathDoesEndWith,
-    RelativePathDoesNotEndWith,
-
-    IRelativePathDoesEndWith,
-    IRelativePathDoesNotEndWith,
-
-    AbsolutePathIs,
-    AbsolutePathIsNot,
-
-    IAbsolutePathIs,
-    IAbsolutePathIsNot,
-
-    AbsolutePathDoesStartWith,
-    AbsolutePathDoesNotStartWith,
-
-    IAbsolutePathDoesStartWith,
-    IAbsolutePathDoesNotStartWith,
-
-    AbsolutePathDoesContain,
-    AbsolutePathDoesNotContain,
-
-    IAbsolutePathDoesContain,
-    IAbsolutePathDoesNotContain,
-
-    AbsolutePathDoesEndWith,
-    AbsolutePathDoesNotEndWith,
-
-    IAbsolutePathDoesEndWith,
-    IAbsolutePathDoesNotEndWith,
-}
-
-impl NodeFilter {
-    fn apply(&self, node: &Node, input: &str) -> bool {
-        match self {
-            Self::RelativePathIs => node.relative_path.eq(input),
-            Self::IRelativePathIs => {
-                node.relative_path.eq_ignore_ascii_case(input)
-            }
-
-            Self::RelativePathIsNot => !node.relative_path.eq(input),
-            Self::IRelativePathIsNot => {
-                !node.relative_path.eq_ignore_ascii_case(input)
-            }
-
-            Self::RelativePathDoesStartWith => {
-                node.relative_path.starts_with(input)
-            }
-            Self::IRelativePathDoesStartWith => node
-                .relative_path
-                .to_lowercase()
-                .starts_with(&input.to_lowercase()),
-
-            Self::RelativePathDoesNotStartWith => {
-                !node.relative_path.starts_with(input)
-            }
-
-            Self::IRelativePathDoesNotStartWith => !node
-                .relative_path
-                .to_lowercase()
-                .starts_with(&input.to_lowercase()),
-
-            Self::RelativePathDoesContain => node.relative_path.contains(input),
-            Self::IRelativePathDoesContain => node
-                .relative_path
-                .to_lowercase()
-                .contains(&input.to_lowercase()),
-
-            Self::RelativePathDoesNotContain => {
-                !node.relative_path.contains(input)
-            }
-            Self::IRelativePathDoesNotContain => !node
-                .relative_path
-                .to_lowercase()
-                .contains(&input.to_lowercase()),
-
-            Self::RelativePathDoesEndWith => {
-                node.relative_path.ends_with(input)
-            }
-            Self::IRelativePathDoesEndWith => node
-                .relative_path
-                .to_lowercase()
-                .ends_with(&input.to_lowercase()),
-
-            Self::RelativePathDoesNotEndWith => {
-                !node.relative_path.ends_with(input)
-            }
-            Self::IRelativePathDoesNotEndWith => !node
-                .relative_path
-                .to_lowercase()
-                .ends_with(&input.to_lowercase()),
-
-            Self::AbsolutePathIs => node.absolute_path.eq(input),
-            Self::IAbsolutePathIs => {
-                node.absolute_path.eq_ignore_ascii_case(input)
-            }
-
-            Self::AbsolutePathIsNot => !node.absolute_path.eq(input),
-            Self::IAbsolutePathIsNot => {
-                !node.absolute_path.eq_ignore_ascii_case(input)
-            }
-
-            Self::AbsolutePathDoesStartWith => {
-                node.absolute_path.starts_with(input)
-            }
-            Self::IAbsolutePathDoesStartWith => node
-                .absolute_path
-                .to_lowercase()
-                .starts_with(&input.to_lowercase()),
-
-            Self::AbsolutePathDoesNotStartWith => {
-                !node.absolute_path.starts_with(input)
-            }
-            Self::IAbsolutePathDoesNotStartWith => !node
-                .absolute_path
-                .to_lowercase()
-                .starts_with(&input.to_lowercase()),
-
-            Self::AbsolutePathDoesContain => node.absolute_path.contains(input),
-            Self::IAbsolutePathDoesContain => node
-                .absolute_path
-                .to_lowercase()
-                .contains(&input.to_lowercase()),
-
-            Self::AbsolutePathDoesNotContain => {
-                !node.absolute_path.contains(input)
-            }
-            Self::IAbsolutePathDoesNotContain => !node
-                .absolute_path
-                .to_lowercase()
-                .contains(&input.to_lowercase()),
-
-            Self::AbsolutePathDoesEndWith => {
-                node.absolute_path.ends_with(input)
-            }
-            Self::IAbsolutePathDoesEndWith => node
-                .absolute_path
-                .to_lowercase()
-                .ends_with(&input.to_lowercase()),
-
-            Self::AbsolutePathDoesNotEndWith => {
-                !node.absolute_path.ends_with(input)
-            }
-            Self::IAbsolutePathDoesNotEndWith => !node
-                .absolute_path
-                .to_lowercase()
-                .ends_with(&input.to_lowercase()),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Eq, Hash, PartialEq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct NodeFilterApplicable {
-    pub filter: NodeFilter,
-    pub input: String,
-}
-
-impl NodeFilterApplicable {
-    pub fn new(filter: NodeFilter, input: String) -> Self {
-        Self { filter, input }
-    }
-
-    fn apply(&self, node: &Node) -> bool {
-        self.filter.apply(node, &self.input)
-    }
-}
-
-#[derive(Debug, Default, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub struct ExplorerConfig {
-    pub filters: IndexSet<NodeFilterApplicable>,
-    pub sorters: IndexSet<NodeSorterApplicable>,
-}
-
-impl ExplorerConfig {
-    pub fn filter(&self, node: &Node) -> bool {
-        self.filters.iter().all(|f| f.apply(node))
-    }
-
-    pub fn sort(&self, a: &Node, b: &Node) -> Ordering {
-        let mut ord = Ordering::Equal;
-        for s in self.sorters.iter() {
-            ord = ord.then(s.apply(a, b));
-        }
-        ord
-    }
-
-    /// Get a reference to the explorer config's filters.
-    pub fn filters(&self) -> &IndexSet<NodeFilterApplicable> {
-        &self.filters
-    }
-
-    /// Get a reference to the explorer config's sorters.
-    pub fn sorters(&self) -> &IndexSet<NodeSorterApplicable> {
-        &self.sorters
-    }
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub enum ExternalMsg {
-    /// Explore the present working directory and register the filtered nodes.
-    /// This operation is expensive. So, try to avoid using it too often.
-    ExplorePwd,
-
-    /// Explore the present working directory and register the filtered nodes asynchronously.
-    /// This operation happens asynchronously. That means, the xplr directory buffers won't be updated
-    /// immediately. Hence, it needs to be used with care and probably with special checks in place.
-    /// To explore `$PWD` synchronously, use `ExplorePwd` instead.
-    ExplorePwdAsync,
-
-    /// Explore the present working directory along with its parents and register the filtered nodes.
-    /// This operation happens asynchronously. That means, the xplr directory buffers won't be updated
-    /// immediately. Hence, it needs to be used with care and probably with special checks in place.
-    /// To explore just the `$PWD` synchronously, use `ExplorePwd` instead.
-    ExploreParentsAsync,
-
-    /// Refresh the UI.
-    /// But it will not re-explore the directory if the working directory is the same.
-    /// If there is some change in the working directory and you want to re-explore it,
-    /// use the `Explore` message instead.
-    /// Also, it will not clear the screen. Use `ClearScreen` for that.
-    Refresh,
-
-    /// Clears the screen.
-    ClearScreen,
-
-    /// Focus next node.
-    FocusNext,
-
-    /// Focus on the `n`th node relative to the current focus where `n` is a given value.
-    ///
-    /// **Example:** `FocusNextByRelativeIndex: 2`
-    FocusNextByRelativeIndex(usize),
-
-    /// Focus on the `n`th node relative to the current focus where `n` is read from
-    /// the input buffer.
-    FocusNextByRelativeIndexFromInput,
-
-    /// Focus on the previous item.
-    FocusPrevious,
-
-    /// Focus on the `-n`th node relative to the current focus where `n` is a given value.
-    ///
-    /// **Example:** `FocusPreviousByRelativeIndex: 2`
-    FocusPreviousByRelativeIndex(usize),
-
-    /// Focus on the `-n`th node relative to the current focus where `n` is read from
-    /// the input buffer.
-    FocusPreviousByRelativeIndexFromInput,
-
-    /// Focus on the first node.
-    FocusFirst,
-
-    /// Focus on the last node.
-    FocusLast,
-
-    /// Focus on the given path.
-    ///
-    /// **Example:** `FocusPath: /tmp`
-    FocusPath(String),
-
-    /// Focus on the path read from input buffer.
-    FocusPathFromInput,
-
-    /// Focus on the absolute `n`th node where `n` is a given value.
-    ///
-    /// **Example:** `FocusByIndex: 2`
-    FocusByIndex(usize),
-
-    /// Focus on the absolute `n`th node where `n` is read from the input buffer.
-    FocusByIndexFromInput,
-
-    /// Focus on the file by name from the present working directory.
-    ///
-    /// **Example:** `FocusByFileName: README.md`
-    FocusByFileName(String),
-
-    /// Change the present working directory ($PWD)
-    ///
-    /// **Example:** `ChangeDirectory: /tmp`
-    ChangeDirectory(String),
-
-    /// Enter into the currently focused path if it's a directory.
-    Enter,
-
-    /// Go back to the parent directory.
-    Back,
-
-    /// Go to the last path visited.
-    LastVisitedPath,
-
-    /// Go to the next path visited.
-    NextVisitedPath,
-
-    /// Follow the symlink under focus to its actual location.
-    FollowSymlink,
-
-    /// Update the input buffer using cursor based operations.
-    ///
-    /// **Example:** `UpdateInputBuffer: GoToPreviousWord`
-    UpdateInputBuffer(InputOperation),
-
-    /// Update the input buffer from given key
-    UpdateInputBufferFromKey,
-
-    /// Append/buffer the given string into the input buffer.
-    ///
-    /// **Example:** `BufferInput: foo`
-    BufferInput(String),
-
-    /// Append/buffer the characted read from a keyboard input into the
-    /// input buffer.
-    BufferInputFromKey,
-
-    /// Set/rewrite the input buffer with the given string.
-    /// When the input buffer is not-null (even if empty string)
-    /// it will show in the UI.
-    ///
-    /// **Example:** `SetInputBuffer: foo`
-    SetInputBuffer(String),
-
-    /// Remove input buffer's last character.
-    RemoveInputBufferLastCharacter,
-
-    /// Remove input buffer's last word.
-    RemoveInputBufferLastWord,
-
-    /// Reset the input buffer back to null. It will not show in the UI.
-    ResetInputBuffer,
-
-    /// Switch input mode.
-    /// It clears the input buffer.
-    ///
-    /// > **NOTE:** To be specific about which mode to switch to, use `SwitchModeBuiltin` or
-    /// `SwitchModeCustom` instead.
-    ///
-    /// **Example:** `SwitchMode: default`
-    SwitchMode(String),
-
-    /// Switch input mode.
-    /// It keeps the input buffer.
-    ///
-    /// > **NOTE:** To be specific about which mode to switch to, use
-    /// `SwitchModeBuiltinKeepingInputBuffer` or
-    /// `SwitchModeCustomKeepingInputBuffer` instead.
-    ///
-    /// **Example:** `SwitchModeKeepingInputBuffer: default`
-    SwitchModeKeepingInputBuffer(String),
-
-    /// Switch to a builtin mode.
-    /// It clears the input buffer.
-    ///
-    /// **Example:** `SwitchModeBuiltin: default`
-    SwitchModeBuiltin(String),
-
-    /// Switch to a builtin mode.
-    /// It keeps the input buffer.
-    ///
-    /// **Example:** `SwitchModeBuiltinKeepingInputBuffer: default`
-    SwitchModeBuiltinKeepingInputBuffer(String),
-
-    /// Switch to a custom mode.
-    /// It clears the input buffer.
-    ///
-    /// **Example:** `SwitchModeCustom: my_custom_mode`
-    SwitchModeCustom(String),
-
-    /// Switch to a custom mode.
-    /// It keeps the input buffer.
-    ///
-    /// **Example:** `SwitchModeCustomKeepingInputBuffer: my_custom_mode`
-    SwitchModeCustomKeepingInputBuffer(String),
-
-    /// Pop the last mode from the history and switch to it.
-    /// It clears the input buffer.
-    PopMode,
-
-    /// Pop the last mode from the history and switch to it.
-    /// It keeps the input buffer.
-    PopModeKeepingInputBuffer,
-
-    /// Switch layout.
-    ///
-    /// > **NOTE:** To be specific about which layout to switch to, use `SwitchLayoutBuiltin` or
-    /// `SwitchLayoutCustom` instead.
-    ///
-    /// **Example:** `SwitchLayout: default`
-    SwitchLayout(String),
-
-    /// Switch to a builtin layout.
-    ///
-    /// **Example:** `SwitchLayoutBuiltin: default`
-    SwitchLayoutBuiltin(String),
-
-    /// Switch to a custom layout.
-    ///
-    /// **Example:** `SwitchLayoutCustom: my_custom_layout`
-    SwitchLayoutCustom(String),
-
-    /// Call a shell command with the given arguments.
-    /// Note that the arguments will be shell-escaped.
-    /// So to read the variables, the `-c` option of the shell
-    /// can be used.
-    /// You may need to pass `ExplorePwd` depening on the expectation.
-    ///
-    /// **Example:** `Call: {command: bash, args: ["-c", "read -p test"]}`
-    Call(Command),
-
-    /// Like `Call` but without the flicker. The stdin, stdout
-    /// stderr will be piped to null. So it's non-interactive.
-    ///
-    /// **Example:** `CallSilently: {command: tput, args: ["bell"]}`
-    CallSilently(Command),
-
-    /// An alias to `Call: {command: bash, args: ["-c", "${command}"], silent: false}`
-    /// where ${command} is the given value.
-    ///
-    /// **Example:** `BashExec: "read -p test"`
-    BashExec(String),
-
-    /// Like `BashExec` but without the flicker. The stdin, stdout
-    /// stderr will be piped to null. So it's non-interactive.
-    ///
-    /// **Example:** `BashExecSilently: "tput bell"`
-    BashExecSilently(String),
-
-    /// Call a Lua function.
-    /// A `CallLuaArg` object will be passed to the function as argument.
-    /// The function can optionally return a list of messages for xplr to handle
-    /// after the executing the function.
-    ///
-    /// **Example:** `CallLua: custom.some_custom_funtion`
-    CallLua(String),
-
-    /// Like `CallLua` but without the flicker. The stdin, stdout
-    /// stderr will be piped to null. So it's non-interactive.
-    ///
-    /// **Example:** `CallLuaSilently: custom.some_custom_function`
-    CallLuaSilently(String),
-
-    /// Execute Lua code without needing to define a function.
-    /// However, `CallLuaArg` won't be available.
-    ///
-    /// **Example:** `LuaEval: "return io.read()"`
-    LuaEval(String),
-
-    /// Like `LuaEval` but without the flicker. The stdin, stdout
-    /// stderr will be piped to null. So it's non-interactive.
-    ///
-    /// **Example:** `LuaEvalSilently: "return { LogInfo = 'foo' }"`
-    LuaEvalSilently(String),
-
-    /// Select the focused node.
-    Select,
-
-    /// Select all the visible nodes.
-    SelectAll,
-
-    /// Select the given path.
-    ///
-    /// **Example:** `SelectPath: "/tmp"`
-    SelectPath(String),
-
-    /// Unselect the focused node.
-    UnSelect,
-
-    /// Unselect all the visible nodes.
-    UnSelectAll,
-
-    /// UnSelect the given path.
-    ///
-    /// **Example:** `UnSelectPath: "/tmp"`
-    UnSelectPath(String),
-
-    /// Toggle selection on the focused node.
-    ToggleSelection,
-
-    /// Toggle between select all and unselect all.
-    ToggleSelectAll,
-
-    /// Toggle selection by file path.
-    ///
-    /// **Example:** `ToggleSelectionByPath: "/tmp"`
-    ToggleSelectionByPath(String),
-
-    /// Clear the selection.
-    ClearSelection,
-
-    /// Add a filter to exclude nodes while exploring directories.
-    ///
-    /// **Example:** `AddNodeFilter: {filter: RelativePathDoesStartWith, input: foo}`
-    AddNodeFilter(NodeFilterApplicable),
-
-    /// Remove an existing filter.
-    ///
-    /// **Example:** `RemoveNodeFilter: {filter: RelativePathDoesStartWith, input: foo}`
-    RemoveNodeFilter(NodeFilterApplicable),
-
-    /// Remove a filter if it exists, else, add a it.
-    ///
-    /// **Example:** `ToggleNodeFilter: {filter: RelativePathDoesStartWith, input: foo}`
-    ToggleNodeFilter(NodeFilterApplicable),
-
-    /// Add a node filter reading the input from the buffer.
-    ///
-    /// **Example:** `AddNodeFilterFromInput: RelativePathDoesStartWith`
-    AddNodeFilterFromInput(NodeFilter),
-
-    /// Remove a node filter reading the input from the buffer.
-    ///
-    /// **Example:** `RemoveNodeFilterFromInput: RelativePathDoesStartWith`
-    RemoveNodeFilterFromInput(NodeFilter),
-
-    /// Remove the last node filter.
-    RemoveLastNodeFilter,
-
-    /// Reset the node filters back to the default configuration.
-    ResetNodeFilters,
-
-    /// Clear all the node filters.
-    ClearNodeFilters,
-
-    /// Add a sorter to sort nodes while exploring directories.
-    ///
-    /// **Example:** `AddNodeSorter: {sorter: ByRelativePath, reverse: false}`
-    AddNodeSorter(NodeSorterApplicable),
-
-    /// Remove an existing sorter.
-    ///
-    /// **Example:** `RemoveNodeSorter: ByRelativePath`
-    RemoveNodeSorter(NodeSorter),
-
-    /// Reverse a node sorter.
-    ///
-    /// **Example:** `ReverseNodeSorter: ByRelativePath`
-    ReverseNodeSorter(NodeSorter),
-
-    /// Remove a sorter if it exists, else, add a it.
-    ///
-    /// **Example:** `ToggleSorterSorter: {sorter: ByRelativePath, reverse: false}`
-    ToggleNodeSorter(NodeSorterApplicable),
-
-    /// Reverse the node sorters.
-    ReverseNodeSorters,
-
-    /// Remove the last node sorter.
-    RemoveLastNodeSorter,
-
-    /// Reset the node sorters back to the default configuration.
-    ResetNodeSorters,
-
-    /// Clear all the node sorters.
-    ClearNodeSorters,
-
-    /// Enable mouse
-    EnableMouse,
-
-    /// Disable mouse
-    DisableMouse,
-
-    /// Toggle mouse
-    ToggleMouse,
-
-    /// Start piping the focused path to the given fifo path
-    ///
-    /// **Example:** `StartFifo: /tmp/xplr.fifo`
-    StartFifo(String),
-
-    /// Close the active fifo and stop piping.
-    StopFifo,
-
-    /// Toggle betwen {Start|Stop}Fifo
-    ToggleFifo(String),
-
-    /// Log information message.
-    ///
-    /// **Example:** `LogInfo: launching satellite`
-    LogInfo(String),
-
-    /// Log a success message.
-    ///
-    /// **Example:** `LogSuccess: satellite reached destination`.
-    LogSuccess(String),
-
-    /// Log an warning message.
-    ///
-    /// **Example:** `LogWarning: satellite is heating`
-    LogWarning(String),
-
-    /// Log an error message.
-    ///
-    /// **Example:** `LogError: satellite crashed`
-    LogError(String),
-
-    /// Quit with returncode zero (success).
-    Quit,
-
-    /// Print $PWD and quit.
-    PrintPwdAndQuit,
-
-    /// Print the path under focus and quit. It can be empty string if there's nothing to focus.
-    PrintFocusPathAndQuit,
-
-    /// Print the selected paths and quit. It can be empty is no path is selected.
-    PrintSelectionAndQuit,
-
-    /// Print the selected paths if it's not empty, else, print the focused node's path.
-    PrintResultAndQuit,
-
-    /// Print the state of application in YAML format. Helpful for debugging or generating
-    /// the default configuration file.
-    PrintAppStateAndQuit,
-
-    /// Write the application state to a file, without quitting. Also helpful for debugging.
-    Debug(String),
-
-    /// Terminate the application with a non-zero return code.
-    Terminate,
-}
-
-impl ExternalMsg {
-    pub fn is_read_only(&self) -> bool {
-        !matches!(
-            self,
-            Self::Call(_)
-                | Self::CallSilently(_)
-                | Self::BashExec(_)
-                | Self::BashExecSilently(_)
-                | Self::CallLua(_)
-                | Self::CallLuaSilently(_)
-                | Self::LuaEval(_)
-                | Self::LuaEvalSilently(_)
-        )
-    }
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub enum MsgIn {
-    Internal(InternalMsg),
-    External(ExternalMsg),
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct Command {
-    pub command: String,
-
-    #[serde(default)]
-    pub args: Vec<String>,
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub enum MsgOut {
-    ExplorePwdAsync,
-    ExploreParentsAsync,
-    Refresh,
-    ClearScreen,
-    Quit,
-    Debug(String),
-    Call(Command),
-    CallSilently(Command),
-    CallLua(String),
-    CallLuaSilently(String),
-    LuaEval(String),
-    LuaEvalSilently(String),
-    Enque(Task),
-    EnableMouse,
-    DisableMouse,
-    ToggleMouse,
-    StartFifo(String),
-    StopFifo,
-    ToggleFifo(String),
-    PrintPwdAndQuit,
-    PrintFocusPathAndQuit,
-    PrintSelectionAndQuit,
-    PrintResultAndQuit,
-    PrintAppStateAndQuit,
-}
 
 #[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
 pub struct Task {
