+++ src/directory_buffer.rs
@@ -1,0 +1,26 @@
+use crate::node::Node;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
+pub struct DirectoryBuffer {
+    pub parent: String,
+    pub nodes: Vec<Node>,
+    pub total: usize,
+    pub focus: usize,
+}
+
+impl DirectoryBuffer {
+    pub fn new(parent: String, nodes: Vec<Node>, focus: usize) -> Self {
+        let total = nodes.len();
+        Self {
+            parent,
+            nodes,
+            total,
+            focus,
+        }
+    }
+
+    pub fn focused_node(&self) -> Option<&Node> {
+        self.nodes.get(self.focus)
+    }
+}
