+++ src/msg/in_/internal.rs
@@ -1,0 +1,10 @@
+use crate::app::DirectoryBuffer;
+use crate::input::Key;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
+pub enum InternalMsg {
+    AddLastFocus(String, Option<String>),
+    SetDirectory(DirectoryBuffer),
+    HandleKey(Key),
+}
